requirements.txt
Django==5.2
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.1
